module_1_basic/requirements.txt
crewai==0.80.0
google-cloud-aiplatform
langchain-google-vertexai==2.0.7
streamlit==1.41.0
python-dotenv==1.0.1
